Fix conversion if the old xp was exactly Integer.MAX_VALUE and use double for getXpNeededToLevelUp (#23)
* Fix conversion if the old xp was exactly Integer.MAX_VALUE
Optimized a bit the conversion and bumped version to 3.6.1

* Use double for getXpNeededToLevelUp
Theoretically speaking it is possible for the EXP needed to level up to be higher than Integer.MAX_VALUE

* Small fix for a parenthesis

* Small formatting fix

* Small formatting fix

+++ src/main/java/me/eccentric_nz/gamemodeinventories/GameModeInventoriesXPCalculator.java
@@ -84,7 +84,7 @@ public class GameModeInventoriesXPCalculator {
      * @return Gets the players current experience points towards the next level.
      */
     public static float getFractionalExp(int level, double exp) {
-        return (float)(exp - getXpForLevel(level))/getXpNeededToLevelUp(level);
+        return (float) ((exp - getXpForLevel(level))/getXpNeededToLevelUp(level));
     }
 
     /**
@@ -113,7 +113,7 @@ public class GameModeInventoriesXPCalculator {
      * @return the amount of experience at this level in the level bar
      * @throws IllegalArgumentException if the level is less than 0
      */
-    public static int getXpNeededToLevelUp(int level) {
+    public static double getXpNeededToLevelUp(int level) {
         Preconditions.checkArgument(level >= 0, "Level may not be negative.");
         return level > 30
             ? 9 * level - 158
